feat(issue-290): split qt mainwindow workspace shell
