+++ modules/qt/include/ui/MainWindowConfig.hpp
@@ -1,0 +1,11 @@
+#ifndef GRAVITY_MODULES_QT_INCLUDE_UI_MAINWINDOWCONFIG_HPP_
+#define GRAVITY_MODULES_QT_INCLUDE_UI_MAINWINDOWCONFIG_HPP_
+
+/*
+ * Module: ui
+ * Responsibility: Mirror the MainWindow configuration application implementation unit.
+ */
+
+#include "ui/MainWindow.hpp"
+
+#endif // GRAVITY_MODULES_QT_INCLUDE_UI_MAINWINDOWCONFIG_HPP_
