+++ modules/qt/include/ui/MainWindowControls.hpp
@@ -1,0 +1,11 @@
+#ifndef GRAVITY_MODULES_QT_INCLUDE_UI_MAINWINDOWCONTROLS_HPP_
+#define GRAVITY_MODULES_QT_INCLUDE_UI_MAINWINDOWCONTROLS_HPP_
+
+/*
+ * Module: ui
+ * Responsibility: Mirror the MainWindow control wiring implementation unit.
+ */
+
+#include "ui/MainWindow.hpp"
+
+#endif // GRAVITY_MODULES_QT_INCLUDE_UI_MAINWINDOWCONTROLS_HPP_
