+++ modules/qt/include/ui/MainWindowFileActions.hpp
@@ -1,0 +1,11 @@
+#ifndef GRAVITY_MODULES_QT_INCLUDE_UI_MAINWINDOWFILEACTIONS_HPP_
+#define GRAVITY_MODULES_QT_INCLUDE_UI_MAINWINDOWFILEACTIONS_HPP_
+
+/*
+ * Module: ui
+ * Responsibility: Mirror the MainWindow file and connector action implementation unit.
+ */
+
+#include "ui/MainWindow.hpp"
+
+#endif // GRAVITY_MODULES_QT_INCLUDE_UI_MAINWINDOWFILEACTIONS_HPP_
