+++ modules/qt/include/ui/MainWindowLayout.hpp
@@ -1,0 +1,11 @@
+#ifndef GRAVITY_MODULES_QT_INCLUDE_UI_MAINWINDOWLAYOUT_HPP_
+#define GRAVITY_MODULES_QT_INCLUDE_UI_MAINWINDOWLAYOUT_HPP_
+
+/*
+ * Module: ui
+ * Responsibility: Mirror the MainWindow layout construction implementation unit.
+ */
+
+#include "ui/MainWindow.hpp"
+
+#endif // GRAVITY_MODULES_QT_INCLUDE_UI_MAINWINDOWLAYOUT_HPP_
