+++ modules/qt/include/ui/MainWindowTelemetry.hpp
@@ -1,0 +1,11 @@
+#ifndef GRAVITY_MODULES_QT_INCLUDE_UI_MAINWINDOWTELEMETRY_HPP_
+#define GRAVITY_MODULES_QT_INCLUDE_UI_MAINWINDOWTELEMETRY_HPP_
+
+/*
+ * Module: ui
+ * Responsibility: Mirror the MainWindow telemetry and tick implementation unit.
+ */
+
+#include "ui/MainWindow.hpp"
+
+#endif // GRAVITY_MODULES_QT_INCLUDE_UI_MAINWINDOWTELEMETRY_HPP_
