+++ modules/qt/include/ui/MainWindowWorkspacePersistence.hpp
@@ -1,0 +1,11 @@
+#ifndef GRAVITY_MODULES_QT_INCLUDE_UI_MAINWINDOWWORKSPACEPERSISTENCE_HPP_
+#define GRAVITY_MODULES_QT_INCLUDE_UI_MAINWINDOWWORKSPACEPERSISTENCE_HPP_
+
+/*
+ * Module: ui
+ * Responsibility: Mirror the MainWindow workspace persistence implementation unit.
+ */
+
+#include "ui/MainWindow.hpp"
+
+#endif // GRAVITY_MODULES_QT_INCLUDE_UI_MAINWINDOWWORKSPACEPERSISTENCE_HPP_
