+++ modules/qt/include/ui/MainWindowWorkspaceShell.hpp
@@ -1,0 +1,11 @@
+#ifndef GRAVITY_MODULES_QT_INCLUDE_UI_MAINWINDOWWORKSPACESHELL_HPP_
+#define GRAVITY_MODULES_QT_INCLUDE_UI_MAINWINDOWWORKSPACESHELL_HPP_
+
+/*
+ * Module: ui
+ * Responsibility: Mirror the MainWindow workspace shell implementation unit.
+ */
+
+#include "ui/MainWindow.hpp"
+
+#endif // GRAVITY_MODULES_QT_INCLUDE_UI_MAINWINDOWWORKSPACESHELL_HPP_
