+++ tests/cmake/targets_gtests.cmake
@@ -222,7 +222,14 @@ if(TARGET Qt6::Widgets AND GRAVITY_TEST_INT_UI_SOURCES)
             "${GRAVITY_ROOT_DIR}/modules/qt/ui/EnergyGraphWidgetPaint.cpp"
             "${GRAVITY_ROOT_DIR}/modules/qt/ui/MainWindowController.cpp"
             "${GRAVITY_ROOT_DIR}/modules/qt/ui/MainWindow.cpp"
+            "${GRAVITY_ROOT_DIR}/modules/qt/ui/MainWindowConfig.cpp"
+            "${GRAVITY_ROOT_DIR}/modules/qt/ui/MainWindowControls.cpp"
+            "${GRAVITY_ROOT_DIR}/modules/qt/ui/MainWindowFileActions.cpp"
+            "${GRAVITY_ROOT_DIR}/modules/qt/ui/MainWindowLayout.cpp"
             "${GRAVITY_ROOT_DIR}/modules/qt/ui/MainWindowPresenter.cpp"
+            "${GRAVITY_ROOT_DIR}/modules/qt/ui/MainWindowTelemetry.cpp"
+            "${GRAVITY_ROOT_DIR}/modules/qt/ui/MainWindowWorkspacePersistence.cpp"
+            "${GRAVITY_ROOT_DIR}/modules/qt/ui/MainWindowWorkspaceShell.cpp"
             "${GRAVITY_ROOT_DIR}/modules/qt/ui/MultiViewWidget.cpp"
             "${GRAVITY_ROOT_DIR}/modules/qt/ui/OctreeOverlay.cpp"
             "${GRAVITY_ROOT_DIR}/modules/qt/ui/OctreeOverlayPainter.cpp"
